requirements.txt
requests
sounddevice
numpy
speechrecognition
pyttsx3
python-dotenv

